Added linux 3.1.1 support to cerowrt
I note that I had to leave about 7 patches out that were mostly
routerstation related, so this version will NOT work on the
routerstation at all. It might fail on other platforms too.

It does work on the wndr3700v2 and 3800.

Also the config file needs some more tweaks to compile cleanly
without requiring you bang on the enter key to generate a new one.

So you will have to do a clean, fresh build, with make V=99
and bang on the enter key to accept the additional kernel
parameters. After you get a kernel built, you can go back
to -j whatever...